requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
werkzeug==3.0.1
